% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairs.R
\name{traverse_pairs}
\alias{traverse_pairs}
\title{Construct source-target pairs}
\usage{
traverse_pairs(sources, targets, method = "all")
}
\arguments{
\item{sources}{Source point \link[terra:SpatVector-class]{terra::SpatVector} or \code{sf} object.}

\item{targets}{Target point \link[terra:SpatVector-class]{terra::SpatVector} or \code{sf} object.}

\item{method}{Pairing method. The initial implementation supports \code{"all"}.}
}
\value{
A data frame with stable pair, source, and target identifiers.
}
\description{
Construct source-target pairs
}
